# Make Scanner a newtype