@Yuras 2016-04-02
Good catch, thank you!

Merged as 19c6649eccdb644a689d6d92ff6f6e3868dc50cc (it required rebase)